Moving calculateSearchFilter to utils
